requirements.txt
appdirs==1.4.4
boto3==1.18.16
botocore==1.21.16
click==8.0.1
dataclasses==0.8
importlib-metadata==4.6.3
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
regex==2021.8.3
s3transfer==0.5.0
six==1.16.0
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
